feat: Dnd 렌더링 최적화
