+++ src/App.tsx
@@ -1,15 +1,12 @@
 import { ChakraProvider } from '@chakra-ui/react';
 import { ScheduleProvider } from './ScheduleContext.tsx';
 import { ScheduleTables } from './ScheduleTables.tsx';
-import ScheduleDndProvider from './ScheduleDndProvider.tsx';
 
 function App() {
 	return (
 		<ChakraProvider>
 			<ScheduleProvider>
-				<ScheduleDndProvider>
-					<ScheduleTables />
-				</ScheduleDndProvider>
+				<ScheduleTables />
 			</ScheduleProvider>
 		</ChakraProvider>
 	);
