+++ src/DraggableSchedule.tsx
@@ -1,0 +1,81 @@
+import React, { ComponentProps } from 'react';
+import {
+	Box,
+	Button,
+	Popover,
+	PopoverArrow,
+	PopoverBody,
+	PopoverCloseButton,
+	PopoverContent,
+	PopoverTrigger,
+	Text,
+} from '@chakra-ui/react';
+import { useDraggable } from '@dnd-kit/core';
+import { CSS } from '@dnd-kit/utilities';
+import { Schedule } from './types.ts';
+import { CellSize, DAY_LABELS } from './constants.ts';
+
+// 단일 강의 아이템
+const DraggableSchedule = React.memo(
+	({
+		id,
+		data,
+		bg,
+		onDeleteButtonClick,
+	}: { id: string; data: Schedule } & ComponentProps<typeof Box> & {
+			onDeleteButtonClick: () => void;
+		}) => {
+		const { day, range, room, lecture } = data;
+
+		const { attributes, setNodeRef, listeners, transform } = useDraggable({ id }); // 특정 아이템을 드래그 가능하게 설정
+		// id - 드래그 아이템 구분
+		// setNodeRef - 드래그할 노드 연결
+		// listeners - 마우스/터치 이벤트 리스너
+		// transform - 현재 드래그 중 이동 위치
+
+		const leftIndex = DAY_LABELS.indexOf(day as (typeof DAY_LABELS)[number]);
+		const topIndex = range[0] - 1;
+		const size = range.length;
+
+		return (
+			<Popover>
+				<PopoverTrigger>
+					<Box
+						position="absolute"
+						left={`${120 + CellSize.WIDTH * leftIndex + 1}px`}
+						top={`${40 + (topIndex * CellSize.HEIGHT + 1)}px`}
+						width={CellSize.WIDTH - 1 + 'px'}
+						height={CellSize.HEIGHT * size - 1 + 'px'}
+						bg={bg}
+						p={1}
+						boxSizing="border-box"
+						cursor="pointer"
+						ref={setNodeRef}
+						transform={CSS.Translate.toString(transform)}
+						{...listeners}
+						{...attributes}
+					>
+						<Text fontSize="sm" fontWeight="bold">
+							{lecture.title}
+						</Text>
+						<Text fontSize="xs">{room}</Text>
+					</Box>
+				</PopoverTrigger>
+
+				{/* 강의 삭제 모달 */}
+				<PopoverContent onClick={(event) => event.stopPropagation()}>
+					<PopoverArrow />
+					<PopoverCloseButton />
+					<PopoverBody>
+						<Text>강의를 삭제하시겠습니까?</Text>
+						<Button colorScheme="red" size="xs" onClick={onDeleteButtonClick}>
+							삭제
+						</Button>
+					</PopoverBody>
+				</PopoverContent>
+			</Popover>
+		);
+	}
+);
+
+export default DraggableSchedule;
